Rework CInButtonState, move cs2/re stuff into its own directory, update RecvPktInfo_t

+++ src/utils/schema.h
@@ -22,8 +22,9 @@
 class CEntityInstance;
 class CBasePlayerController;
 
-struct SchemaKey {
-	int16_t offset;
+struct SchemaKey
+{
+	int32 offset;
 	bool networked;
 };
